apps/api/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy[asyncio]>=2.0.25,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
celery[redis]>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0
python-multipart>=0.0.6,<1.0.0
pgvector>=0.2.4,<1.0.0
sentence-transformers>=2.3.0,<3.0.0
easyocr>=1.7.0,<2.0.0
faster-whisper>=1.0.0,<2.0.0
opencv-python-headless>=4.9.0,<5.0.0
Pillow>=10.2.0,<11.0.0
imagehash>=4.3.1,<5.0.0
numpy>=1.26.0,<3.0.0
scipy>=1.12.0,<2.0.0
pandas>=2.2.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.27.0,<1.0.0
